+++ clients/privacy-center/components/RequestModal.tsx
@@ -91,6 +91,7 @@ const useRequestForm = ({
           headers: {
             Accept: "application/json",
             "Content-Type": "application/json",
+            "X-Fides-Source": "fidesops-privacy-center",
           },
           body: JSON.stringify(body),
         });