+++ src/fidesops/task/graph_task.py
@@ -63,9 +63,6 @@ def retry(
             self = args[0]
 
             raised_ex = None
-            if config.dev_mode:
-                # If dev mode, return here so exception isn't caught
-                return func(*args, **kwargs)
             for attempt in range(config.execution.TASK_RETRY_COUNT + 1):
                 try:
                     self.skip_if_disabled()